[REFACTOR] simplification of the SRFBA implementation (removing linear mixin)
