+++ src/mewpy/mew/lp/__init__.py
@@ -1,3 +1,3 @@
-from mewpy.mew.lp.linear_problem import LinearProblem
-from mewpy.mew.lp.linear_mixin import LinearMixIn
-from mewpy.mew.lp.linear_containers import VariableContainer, ConstraintContainer
+from .linear_problem import LinearProblem
+from .linear_containers import VariableContainer, ConstraintContainer, concat_constraints
+from .linear_utils import integer_coefficients
